WIP cassandra-test simple cluster
Deploys a small 4-node cluster and runs nft job against it.
The nft job uses cassandra-stress to pre-load the data, warmup the JVM
(cassandra-stress defaults) and actually run the load.
Not much thought has been given to the number or table design,
so consider the load / design WIP.
Here we're more interested in getting a basic setup in place to be able
to run a load tests.