Reconcile by determining current state (#170)
This is a follow up to #150 and supercedes #164
This changes the reconciliation logic to reconstruct the Cassandra definition from the current state instead of tracking the previous resources state in memory which doesn't survive operator redeployments.

This approach allows for resources to be reconciled when the operator is restarted and _eventually_ when the resources are modified outside of the control of the operator (e.g. editing a statefulset, unintentionally deleted a job). I say _eventually_ because at the moment each change is applied asynchronously via our `dispatcher`. As a result, changes that fail the first time may not be applied before the next reconciliation loop (either at next change or at worst at the next `manager.Options.SyncPeriod` which defaults to 10hrs). Follow-up work will be done to address that, perhaps by swapping event dispatching and reconciliation, reducing the `SyncPeriod`, improving the changes application or a combination of that. 

As part of this change, the list of clusters is now managed by the `reconciler`
The list of clusters is mainly used to allow retrieval of metrics from
clusters managed by the operator.
Previously new clusters where only added on creation and on updates
which did not work for scenarios where the operator was restarted
but nothing needed to be reconciled.
This changes the logic to register new cluster definition at each
reconciliation, so the list is correct even when nothing is changed.

Required by #53