Reconcile partial and external modification (#171)
* Reconcile external modifications

Reconcile the Cassandra cluster headless service, so it gets recreated when unintentionally deleted.
The service creation has been split out of the ADD_CLUSTER operation to make it possible to reconcile statefulset and service independently.
Adds tests showing that Kubernetes resources owned by a Cassandra object are recreated by the operator when they get unintentionally deleted - this includes statefulset, headless service and cronjobs.

* Reconcile partial updates

Adds tests showing partial updates are also reconciled by the operator. For instance creating racks that were not initially created and rolling out custom config to racks that were not updated. Partial updates may happen when the operator is stopped half-way through applying changes or when the API server isn't responding.

* Exits reconciliation early when Cassandra is being deleted

If Cassandra is deleted with the default cascading policy,
it will be marked with a DeletionTimestamp while its child resources are
deleted. As the deletion of the resources can take some time,
we exit the reconcile loop right away to make sure the reconciler does
not attempt to reconcile resurrect child resources being deleted.

* Skip event processing when Cassandra deleted

Do not attempt to update the cluster when it is deleted or in the
process of being deleted.
Due to the asynchronous nature of the event processing, it is possible
to have pending events that have yet to be applied even though the
cluster has been (or is being) deleted. So to prevent resurrecting the
cluster we skip all modifications.