Change CassandraPod spec to use ResourceRequirements (#178)
* Make pod use ResourceRequirements

This is a breaking API change.
It changes the CassandraPod specs to use `core.v1.ResourceRequirements` Kubernetes type instead of the single value we use for cpu and memory. This effectively allows for both requests and limits to be specified for cpu and memory. Cpu requests and limits continue to be optional and memory requests and limits mandatory. When both requests and limits are provided, we validate that requests <= limits.

For #79 

* Allow removal of cpu request and limit

This is to cover the scenario where a cpu request or limit is never set
or initially set but later removed. Setting a request or limit to "0"
is not the same as not setting it at all; when not setting a value, the
value is unbounded.
Besides kubernetes resources policy will reject resources changes when
the limit is lower than the request - e.g. you cannot set a request to
10Mi with a limit at 0, but removing the limit altogether is fine.

Memory request and limit continues to be mandatory, so no need to
handle the value removal scenarios