Fix pod cpu resources removal by updating rather than patching StatefulSet (#182)
When processing a cluster update we used to create a json patch with the delta and apply it using the `patch` verb for each `StatefulSet` that needed changing.
As we used a strategic json merge patch strategy this worked well as long as we needed to update existing fields or add new values. This however didn't work so well as shown in #181 when attempting to remove properties or elements from a list.

This PR changes the logic to load the current `StatefulSet` object and update it to bring it to the desired state. This removes the need to "determine" the changes delta and therefore simplifies the code. As we update the entire `StatefulSetSpec` when updating the object based on the desired state, the resulting `StatefulSet` has similar properties than if it was newly created. Any additional labels, annotations or other fields that might have been added outside of the operator control would therefore be wiped out. 

Fix #181