Flexible storage type per rack (#184)
This overhauls the way we define rack storage to add flexibility to the Cassandra specification by allowing different types of storage and eventually multiple storages per rack. 
We envisage that most users would want to have at least `/var/logs/cassandra` and `/var/lib/cassandra` mounted in different volumes, with the first an `emptyDir` and the second a `PV`. 
This PR intentionally restricts the number of custom storage per rack to 1 to reduce the scope of the changes. Support for multiple storage per rack will be added in a follow-up PR.

We support `persistentVolumeClaim` and `emptyDir` volume types as first class citizen.
Inf fact the API exposes `v1.PersistentVolumeClaimSpec` and `v1.EmptyDirVolumeSource` objects respectively, which gives full control on the storage properties and allows our API to evolve alongside Kubernetes. 

`storageClassName` is no longer required as specified in Kubernetes persistent volume definition - see [Class](https://kubernetes.io/docs/concepts/storage/persistent-volumes/#class-1) 

This is a breaking change as it moves the storage size definition from the Pod to the Rack along side the storage class.

Required by #78