Allow multiple rack storages (#190)
This follows up on #184 to allow creating clusters with multiple custom volumes per rack.
This is so we can support custom volume mounts with different volume types. For instance we envisage that most users would want to have at least 2 volumes to separate logs from data and so have `/var/log/cassandra` and `/var/lib/cassandra` mounted in different volumes, with perhaps the first an `emptyDir` and the second a `PV`. In which case the Cassandra definition would look like this:
```
    racks:
    - name: a
      replicas: 1
      storage:
      - path: "/var/lib/cassandra"
        persistentVolumeClaim:
          resources:
            requests:
              storage: 1Gi
          storageClassName: standard-zone-a
      - path: "/var/log/cassandra"
        emptyDir: {}        
      zone: eu-west-1a
```

When using multiple volumes the `path` must be unique.
It is also used to derive the name of the volume and associated mount in the container, by stripping the lead `/` and replacing remaining `/` with `-`  e.g. a volume mount with path: `/var/lib/cassandra` would have a volume named `var-lib-cassandra`

Custom volumes can be defined when creating the cluster initially, but can also be added at a later stage by editing the Cassandra definition. Volume removal and volume property updates are not supported. 

Required by #78