Increase event publication to 1min (#200)
As due to the async nature of the event processing, the operator will
end up processing more events than necessary as each change to a watched
resource triggers a new reconciliation. Each reconciliation creates a new set of
events even though changes for previous events are still being applied. 
This is a known issue that is addressed as part of #173.
In the meantime give more time for tests to see the events publication.