Run containers as non-root users (#208)
Up until now, the operator and all containers started on the provisioned Cassandra pods ran as the root user. This meant that the operator wasn't usable by default within Kubernetes deployments which forbid containers running as the root user.

Part of this restriction was due to use using phusion/baseimage as the base image for the operator and sidecar images, as that requires root-level access to perform certain operations. I've switched these images over to using alpine instead, which solves that problem and also gives us the benefit of a reduced image size.

In order to make these changes testable, I've modified the kind cluster config so that a permissive pod security policy is used in kube-system and a restrictive one is used in test-cassandra-operator.

Fixes #206
Fixes #207