Implement operation interruption (#212)
Introduced a new controller, the Interrupter. The purpose of this is to spot when a cluster definition has changed, check whether the Reconciler is waiting for a statefulset creation from a previous reconciliation attempt to complete, and if so, interrupt the Reconciler.

This has been implemented through a shared synchronised map of ActiveReconciliation structs. Each of these records the revision number of the Cassandra and ConfigMap being used for a revision, along with a channel which can be used by the Interrupter to communicate with the Reconciler. The map is populated with new ActiveReconciliations by the Reconciler, which closes the
channel and removes the record from the map when the reconciliation completes.

In the Reconciler's statefulset wait loop, it checks on every poll whether any message has been received on the channel, and if so, it exits the polling by returning an ErrReconciliationInterrupted. This is then used by the various operations to signify that the reconciliation attempt should not progress any further.

On the Interrupter side, it checks the Cassandra and ConfigMap versions in the ActiveReconciliation, and if either of these differs from the current versions fetched from the API server, it will send a message down the channel to effect the interruption.

Fixes #180