Improve robustness of E2E tests (#224)
By default, the E2E tests run against a Kind cluster which is created and configured by supplied build scripts, but there may be cases where one wants to deploy the operator into an existing cluster and run the E2E tests against it there.

This PR fixes a couple of issues which may arise when testing the operator in another cluster:

There may be certain groups of tests which should be skipped. If so, these can be supplied as a comma-separated list of relative package paths in the SKIP_PACKAGES environment variable, e.g. SKIP_PACKAGES=test/e2e/parallel/validation,test/e2e/parallel/modification.
Clusters may be running with LimitRanges defined in namespaces, in which case operator-created pods may have non-nil CPU requests and limits in the event that the input cassandra resource left these undefined.
Clusters may be running on different platforms or with different versions of Docker, which will emit different error messages when attempting to pull an image that doesn't exist.